requirements.txt
bcrypt==4.0.1
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==2.3.0
Flask-Cors==4.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==8.0.33
protobuf==3.20.3
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==3.1.3
